+++ src/AgentNotify.Api/WebUi/wwwroot/js/views/router-settings.js
@@ -1,7 +1,0 @@
-import { renderRouter } from "./router-shared.js";
-
-export default {
-  async render(page, ctx) {
-    await renderRouter(page, ctx, "settings");
-  },
-};